# Optimize UTF-16 validation code