--- a/src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf16Utility.Validation.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf16Utility.Validation.cs
@@ -60,6 +60,7 @@ namespace System.Text.Unicode
 
             long tempUtf8CodeUnitCountAdjustment = 0;
             int tempScalarCountAdjustment = 0;
+            char* pEndOfInputBuffer = &pInputBuffer[(uint)inputLength];
 
             // Per https://github.com/dotnet/runtime/issues/41699, temporarily disabling
             // ARM64-intrinsicified code paths. ARM64 platforms may still use the vectorized
@@ -69,14 +70,16 @@ namespace System.Text.Unicode
             {
                 if (inputLength >= Vector128<ushort>.Count)
                 {
-                    Vector128<ushort> vector0080 = Vector128.Create((ushort)0x80);
-                    Vector128<ushort> vectorA800 = Vector128.Create((ushort)0xA800);
-                    Vector128<short> vector8800 = Vector128.Create(unchecked((short)0x8800));
-                    Vector128<ushort> vectorZero = Vector128<ushort>.Zero;
+                    Vector128<ushort> vector0080 = Vector128.Create((ushort)0x0080);
+                    Vector128<ushort> vector7800 = Vector128.Create((ushort)0x7800);
+                    Vector128<ushort> vectorA000 = Vector128.Create((ushort)0xA000);
 
                     Vector128<byte> bitMask128 = BitConverter.IsLittleEndian ?
                         Vector128.Create(0x80402010_08040201).AsByte() :
                         Vector128.Create(0x01020408_10204080).AsByte();
+
+                    char* pHighestAddressWhereCanReadOneVector = pEndOfInputBuffer - Vector128<ushort>.Count;
+                    Debug.Assert(pHighestAddressWhereCanReadOneVector >= pInputBuffer);
 
                     do
                     {
@@ -90,6 +93,7 @@ namespace System.Text.Unicode
                             utf16Data = Sse2.LoadVector128((ushort*)pInputBuffer); // unaligned
                         }
 
+                        pInputBuffer += Vector128<ushort>.Count; // eagerly bump this now in preparation for next loop, will adjust later if necessary
                         Vector128<ushort> charIsNonAscii;
 
                         if (AdvSimd.Arm64.IsSupported)
@@ -133,14 +137,20 @@ namespace System.Text.Unicode
                         Vector128<ushort> charIsThreeByteUtf8Encoded;
                         uint mask;
 
+                        // Since 3-byte elements have a value >= 0x0800, we'll perform a saturating add of 0x7800 in order to
+                        // get all 3-byte elements to have their 0x8000 bits set. A saturating add will not set the 0x8000
+                        // bit for 1-byte or 2-byte elements. 2-byte and 3-byte elements MAY have their 0x0080 bits set,
+                        // but this is not required, and the 0x0080 bits (see 'charIsNonAscii' above) will be explicitly
+                        // set for such elements anyway by the immediately following OR instruction.
+
                         if (AdvSimd.IsSupported)
                         {
-                            charIsThreeByteUtf8Encoded = AdvSimd.Subtract(vectorZero, AdvSimd.ShiftRightLogical(utf16Data, 11));
+                            charIsThreeByteUtf8Encoded = AdvSimd.AddSaturate(utf16Data, vector7800);
                             mask = GetNonAsciiBytes(AdvSimd.Or(charIsNonAscii, charIsThreeByteUtf8Encoded).AsByte(), bitMask128);
                         }
                         else
                         {
-                            charIsThreeByteUtf8Encoded = Sse2.Subtract(vectorZero, Sse2.ShiftRightLogical(utf16Data, 11));
+                            charIsThreeByteUtf8Encoded = Sse2.AddSaturate(utf16Data, vector7800);
                             mask = (uint)Sse2.MoveMask(Sse2.Or(charIsNonAscii, charIsThreeByteUtf8Encoded).AsByte());
                         }
 
@@ -168,24 +178,43 @@ namespace System.Text.Unicode
                         // unpaired surrogates in our data. (Unpaired surrogates would invalidate
                         // our computed result and we'd have to throw it away.)
 
-                        uint popcnt = (uint)BitOperations.PopCount(mask);
+                        nuint popcnt = (uint)BitOperations.PopCount(mask); // on x64, perform zero-extension for free
 
                         // Surrogates need to be special-cased for two reasons: (a) we need
                         // to account for the fact that we over-counted in the addition above;
                         // and (b) they require separate validation.
+                        //
+                        // Since surrogate code points are [D800..DFFF], adding {A000} to each element moves surrogate
+                        // code points to [7800..7FFF], which allows performing a single signed comparison.
                         if (AdvSimd.Arm64.IsSupported)
                         {
-                            utf16Data = AdvSimd.Add(utf16Data, vectorA800);
-                            mask = GetNonAsciiBytes(AdvSimd.CompareLessThan(utf16Data.AsInt16(), vector8800).AsByte(), bitMask128);
+                            mask = GetNonAsciiBytes(AdvSimd.CompareLessThan(AdvSimd.Add(utf16Data, vectorA000).AsInt16(), vector7800.AsInt16()).AsByte(), bitMask128);
                         }
                         else
                         {
-                            utf16Data = Sse2.Add(utf16Data, vectorA800);
-                            mask = (uint)Sse2.MoveMask(Sse2.CompareLessThan(utf16Data.AsInt16(), vector8800).AsByte());
+                            mask = (uint)Sse2.MoveMask(Sse2.CompareLessThan(Sse2.Add(utf16Data, vectorA000).AsInt16(), vector7800.AsInt16()).AsByte());
                         }
 
-                        if (mask != 0)
+                    FinishIteration:
+
+                        // Note: mask bits are set when the corresponding element is NOT a surrogate.
+                        // We'll invert this before entering the "validate surrogate pairs" logic below.
+
+                        if (mask == 0xFFFF)
                         {
+                            // Put this logic up top since it's predicted-taken (surrogate pairs are uncommon).
+                            // Either we saw no surrogates or we already handled them below.
+
+                            tempUtf8CodeUnitCountAdjustment += (long)popcnt;
+                            if (pInputBuffer > pHighestAddressWhereCanReadOneVector)
+                            {
+                                goto NonVectorizedLoop; // can no longer read a vector's worth of data
+                            }
+                        }
+                        else
+                        {
+                            mask = ~mask;
+
                             // There's at least one UTF-16 surrogate code unit present.
                             // Since we performed a pmovmskb operation on the result of a 16-bit pcmpgtw,
                             // the resulting bits of 'mask' will occur in pairs:
@@ -194,10 +223,8 @@ namespace System.Text.Unicode
                             //
                             // A UTF-16 high/low surrogate code unit has the bit pattern [ 11011q## ######## ],
                             // where # is any bit; q = 0 represents a high surrogate, and q = 1 represents
-                            // a low surrogate. Since we added 0xA800 in the vectorized operation above,
-                            // our surrogate pairs will now have the bit pattern [ 10000q## ######## ].
-                            // If we logical right-shift each word by 3, we'll end up with the bit pattern
-                            // [ 00010000 q####### ], which means that we can immediately use pmovmskb to
+                            // a low surrogate. Right-shifting each surrogate char by 3 bits, we end up with
+                            // [ 00011011 q####### ], which means that we can immediately use pmovmskb to
                             // determine whether a given char was a high or a low surrogate.
                             //
                             // Therefore the resulting bits of 'mask2' will occur in pairs:
@@ -244,7 +271,7 @@ namespace System.Text.Unicode
                             highSurrogatesMask <<= 2;
                             if ((ushort)highSurrogatesMask != lowSurrogatesMask)
                             {
-                                goto NonVectorizedLoop; // error: mismatched surrogate pair; break out of vectorized logic
+                                break; // error: mismatched surrogate pair; break out of vectorized logic
                             }
 
                             if (highSurrogatesMask > ushort.MaxValue)
@@ -254,8 +281,7 @@ namespace System.Text.Unicode
 
                                 highSurrogatesMask = (ushort)highSurrogatesMask; // don't allow stray high surrogate to be consumed by popcnt
                                 popcnt -= 2; // the '0xC000_0000' bits in the original mask are shifted out and discarded, so account for that here
-                                pInputBuffer--;
-                                inputLength++;
+                                pInputBuffer--; // don't consume this char (pointer has already been bumped at start of loop)
                             }
 
                             // If we're 64-bit, we can perform the zero-extension of the surrogate pairs count for
@@ -286,12 +312,16 @@ namespace System.Text.Unicode
                                 // Take the hit of the 64-bit extension now.
                                 tempUtf8CodeUnitCountAdjustment -= 2 * (uint)surrogatePairsCountNuint;
                             }
-                        }
 
-                        tempUtf8CodeUnitCountAdjustment += popcnt;
-                        pInputBuffer += Vector128<ushort>.Count;
-                        inputLength -= Vector128<ushort>.Count;
-                    } while (inputLength >= Vector128<ushort>.Count);
+                            mask = 0xFFFF; // mark "no surrogates require processing"
+                            goto FinishIteration; // jump backward to continue the main loop
+                        }
+                    } while (true);
+
+                    // If we reached this point, we saw truly invalid data within the loop.
+                    // Need to undo the eager "bump pInputBuffer" adjustment that took place at start of loop.
+
+                    pInputBuffer -= Vector128<ushort>.Count;
                 }
             }
             else if (Vector.IsHardwareAccelerated)
@@ -302,6 +332,9 @@ namespace System.Text.Unicode
                     Vector<ushort> vector0400 = new Vector<ushort>(0x0400);
                     Vector<ushort> vector0800 = new Vector<ushort>(0x0800);
                     Vector<ushort> vectorD800 = new Vector<ushort>(0xD800);
+
+                    char* pHighestAddressWhereCanReadOneVector = pEndOfInputBuffer - Vector<ushort>.Count;
+                    Debug.Assert(pHighestAddressWhereCanReadOneVector >= pInputBuffer);
 
                     do
                     {
@@ -387,7 +420,6 @@ namespace System.Text.Unicode
                                 // We'll adjust our counters so that we don't consider this char consumed.
 
                                 pInputBuffer--;
-                                inputLength++;
                                 popcnt32 -= 2;
                             }
 
@@ -408,8 +440,7 @@ namespace System.Text.Unicode
 
                         tempUtf8CodeUnitCountAdjustment += popcnt32;
                         pInputBuffer += Vector<ushort>.Count;
-                        inputLength -= Vector<ushort>.Count;
-                    } while (inputLength >= Vector<ushort>.Count);
+                    } while (pInputBuffer <= pHighestAddressWhereCanReadOneVector);
                 }
             }
 
@@ -419,7 +450,7 @@ namespace System.Text.Unicode
             // from vectorization, or we saw invalid UTF-16 data in the vectorized code paths and need to
             // drain remaining valid chars before we report failure.
 
-            for (; inputLength > 0; pInputBuffer++, inputLength--)
+            for (; pInputBuffer < pEndOfInputBuffer; pInputBuffer++)
             {
                 uint thisChar = pInputBuffer[0];
                 if (thisChar <= 0x7F)
@@ -444,7 +475,7 @@ namespace System.Text.Unicode
 
                 tempUtf8CodeUnitCountAdjustment -= 2;
 
-                if (inputLength == 1)
+                if ((nuint)pEndOfInputBuffer - (nuint)pInputBuffer < sizeof(uint))
                 {
                     goto Error; // input buffer too small to read a surrogate pair
                 }
@@ -459,7 +490,6 @@ namespace System.Text.Unicode
                 tempUtf8CodeUnitCountAdjustment += 2; // 2 UTF-16 code units -> 4 UTF-8 code units
 
                 pInputBuffer++; // consumed one extra char
-                inputLength--;
             }
 
         Error: