This improves the performance of `Utf8Encoding.GetByteCount(string)` and similar APIs. There are three main optimizations:

1. Reorder the loop logic slightly to keep common paths (no surrogate pairs) together.
2. Reduce the total number of SIMD operations that take place within the main loop.
3. Simplify bookkeeping by bumping just the _pInputBuffer_ pointer without also toggling _inputLength_ within the hot loop.

Additionally, on x64, this reduces the total number of XMM registers used from 7 to 6, which means that the JIT can use only volatile registers (xmm0 - xmm5) throughout the method. This saves a little bit of space in the method prolog + epilog where we previously needed to save + restore xmm6.

__Benchmark results__, using the Project Gutenberg corpus (see also [perf repo sample texts](https://github.com/dotnet/performance/tree/main/src/benchmarks/micro/libraries/Common)):

|       Method |        Job |    Toolchain |      Corpus |      Mean |     Error |    StdDev | Ratio |
|------------- |----------- |------------- |------------ |----------:|----------:|----------:|------:|
| **GetByteCount** | **Job-OJQKGR** |         **main** |    **11-0.txt** | **19.889 μs** | **0.0923 μs** | **0.0771 μs** |  **1.00** |
| GetByteCount | Job-DHGKIB | utf16trans_b |    11-0.txt | 14.748 μs | 0.1517 μs | 0.1419 μs |  0.74 |
|              |            |              |             |           |           |           |       |
| **GetByteCount** | **Job-OJQKGR** |         **main** |      **11.txt** |  **4.977 μs** | **0.0580 μs** | **0.0514 μs** |  **1.00** |
| GetByteCount | Job-DHGKIB | utf16trans_b |      11.txt |  3.849 μs | 0.0276 μs | 0.0230 μs |  0.77 |
|              |            |              |             |           |           |           |       |
| **GetByteCount** | **Job-OJQKGR** |         **main** | **25249-0.txt** | **11.221 μs** | **0.0389 μs** | **0.0345 μs** |  **1.00** |
| GetByteCount | Job-DHGKIB | utf16trans_b | 25249-0.txt |  6.858 μs | 0.0340 μs | 0.0266 μs |  0.61 |
|              |            |              |             |           |           |           |       |
| **GetByteCount** | **Job-OJQKGR** |         **main** | **30774-0.txt** |  **9.593 μs** | **0.0535 μs** | **0.0417 μs** |  **1.00** |
| GetByteCount | Job-DHGKIB | utf16trans_b | 30774-0.txt |  5.788 μs | 0.0231 μs | 0.0216 μs |  0.60 |
|              |            |              |             |           |           |           |       |
| **GetByteCount** | **Job-OJQKGR** |         **main** | **39251-0.txt** | **12.239 μs** | **0.1219 μs** | **0.0952 μs** |  **1.00** |
| GetByteCount | Job-DHGKIB | utf16trans_b | 39251-0.txt |  7.387 μs | 0.0253 μs | 0.0224 μs |  0.60 |

Here's the method hot loop, annotated:

```asm
; Do we even have enough data to warrant vectorization?
00007ff9`fca57068 83ff08           cmp     edi, 8
00007ff9`fca5706b 0f8c0c010000     jl      <dont_vectorize>

; Setup: load frequently used consts into XMM registers
00007ff9`fca57071 c5f9100527010000 vmovupd xmm0, xmmword ptr [System_Private_CoreLib!System.Text.Unicode.Utf16Utility.GetPointerToFirstInvalidChar(Char*,  Int32,  Int64 ByRef,  Int32 ByRef)+0x180 (00007ff9`fca571a0)]
00007ff9`fca57079 c5f9100d2f010000 vmovupd xmm1, xmmword ptr [System_Private_CoreLib!System.Text.Unicode.Utf16Utility.GetPointerToFirstInvalidChar(Char*,  Int32,  Int64 ByRef,  Int32 ByRef)+0x190 (00007ff9`fca571b0)]
00007ff9`fca57081 c5f9101537010000 vmovupd xmm2, xmmword ptr [System_Private_CoreLib!System.Text.Unicode.Utf16Utility.GetPointerToFirstInvalidChar(Char*,  Int32,  Int64 ByRef,  Int32 ByRef)+0x1a0 (00007ff9`fca571c0)]
00007ff9`fca57089 4c8d41f0         lea     r8, [rcx-10h]

; Main loop: read 128 bits and check for valid UTF-16 while also performing UTF-8 / UTF-32 analysis over the text
00007ff9`fca5708d c5fa6f1e         vmovdqu xmm3, xmmword ptr [rsi]
00007ff9`fca57091 4883c610         add     rsi, 10h
00007ff9`fca57095 c4e2613ae0       vpminuw xmm4, xmm3, xmm0 ; process 2-byte chars
00007ff9`fca5709a c5e1dde9         vpaddusw xmm5, xmm3, xmm1 ; process 3-byte chars
00007ff9`fca5709e c5d9ebe5         vpor    xmm4, xmm4, xmm5
00007ff9`fca570a2 c579d7cc         vpmovmskb r9d, xmm4
00007ff9`fca570a6 f3450fb8c9       popcnt  r9d, r9d ; weird: JIT doesn't try to account for popcnt false dependency here?
00007ff9`fca570ab c5e1fde2         vpaddw  xmm4, xmm3, xmm2 ; look for surrogates
00007ff9`fca570af c5f165e4         vpcmpgtw xmm4, xmm1, xmm4
00007ff9`fca570b3 c579d7d4         vpmovmskb r10d, xmm4
00007ff9`fca570b7 4181faffff0000   cmp     r10d, 0FFFFh ; were any surrogates found?
00007ff9`fca570be 750d             jne     <surrogate_chars_found> ; perform some extra validation outside the hot loop
00007ff9`fca570c0 4903c1           add     rax, r9
00007ff9`fca570c3 493bf0           cmp     rsi, r8
00007ff9`fca570c6 76c5             jbe     System_Private_CoreLib!System.Text.Unicode.Utf16Utility.GetPointerToFirstInvalidChar(Char*,  Int32,  Int64 ByRef,  Int32 ByRef)+0x6d (00007ff9`fca5708d) ; start of loop

; Running out of data - go down non-vectorized code
00007ff9`fca570c8 e9b0000000       jmp     <dont_vectorize>
```