@GrabYourPitchforks 2021-04-22
Note to reviewers: The tests in https://github.com/dotnet/runtime/blob/main/src/libraries/System.Runtime/tests/System/Text/Unicode/Utf16UtilityTests.ValidateChars.cs should do a good job of exercising these code paths and the edge cases they need to handle. I don't think we need to add any extra tests beyond those that already exist.