@DrewScoggins 2021-05-05
### Run Information

Architecture | x64
-- | --
OS | ubuntu 18.04
Baseline | [508e56025f66f5c2c00a8f1902816959bb61e0ac](https://github.com/dotnet/runtime/commit/508e56025f66f5c2c00a8f1902816959bb61e0ac)
Compare | [65bcecebde4aebd5926952ccfc4dc8e7fbf7eaed](https://github.com/dotnet/runtime/commit/65bcecebde4aebd5926952ccfc4dc8e7fbf7eaed)
Diff | [Diff](https://github.com/dotnet/runtime/compare/508e56025f66f5c2c00a8f1902816959bb61e0ac...65bcecebde4aebd5926952ccfc4dc8e7fbf7eaed)


### Improvemnts in System.Text.Perf_Utf8Encoding

Benchmark | Baseline | Test | Test/Base | Test Quality | Baseline IR | Compare IR | IR Ratio | Baseline ETL | Compare ETL
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
[GetString - Duration of single invocation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Text.Perf_Utf8Encoding.GetString(Input%3a%20Chinese).html>) | 231.13 μs | 188.57 μs | 0.82 | 0.25 | | |
[GetByteCount - Duration of single invocation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Text.Perf_Utf8Encoding.GetByteCount(Input%3a%20Greek).html>) | 9.89 μs | 8.91 μs | 0.90 | 0.02 | | |
[GetByteCount - Duration of single invocation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Text.Perf_Utf8Encoding.GetByteCount(Input%3a%20EnglishMostlyAscii).html>) | 20.76 μs | 19.35 μs | 0.93 | 0.05 | | |
[GetByteCount - Duration of single invocation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Text.Perf_Utf8Encoding.GetByteCount(Input%3a%20Cyrillic).html>) | 7.87 μs | 7.08 μs | 0.90 | 0.01 | | |
[GetByteCount - Duration of single invocation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Text.Perf_Utf8Encoding.GetByteCount(Input%3a%20Chinese).html>) | 8.99 μs | 8.15 μs | 0.91 | 0.01 | | |

![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding_1.png>)
![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding_2.png>)
![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding_3.png>)
![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding_4.png>)
![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding_5.png>)
[Historical Data in Reporting System](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_26_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Text.Perf_Utf8Encoding.html>)

### Repro

```cmd
git clone https://github.com/dotnet/performance.git
python3 .\performance\scripts\benchmarks_ci.py -f netcoreapp5.0 --filter 'System.Text.Perf_Utf8Encoding*'
```
<details>

### Payloads

[Baseline](<https://helixdi107v0xdeko0k025g8.blob.core.windows.net/helix-job-92d8e263-192b-47a0-b0fc-f8e4663a7132d09c7450b0549b1b2/1e2ad9f0-c937-4f2c-9688-188d1722bc26.zip?sv=2019-07-07&se=2021-05-20T19%3A49%3A56Z&sr=c&sp=rl&sig=aZ5UfiTXxlahVaZNkLNcwOBD6Kl0uNtmCtuZQUU7eXk%3D>)
[Compare](<https://helixdi107v0xdeko0k025g8.blob.core.windows.net/helix-job-e75ce71e-be15-43a4-8c4e-493d688ea41a1177a1f74b64d6ea5/5cf6e11a-0934-4d76-8207-2c96c5667758.zip?sv=2019-07-07&se=2021-05-21T11%3A31%3A07Z&sr=c&sp=rl&sig=61hIA8%2BKm%2FoP6aoNVdJtuEVpDHpuoQC90ER0Mso%2FNhc%3D>)

### Histogram
#### System.Text.Perf_Utf8Encoding.GetString(Input: Chinese)

```log

```
#### System.Text.Perf_Utf8Encoding.GetByteCount(Input: Greek)

```log

```
#### System.Text.Perf_Utf8Encoding.GetByteCount(Input: EnglishMostlyAscii)

```log

```
#### System.Text.Perf_Utf8Encoding.GetByteCount(Input: Cyrillic)

```log

```
#### System.Text.Perf_Utf8Encoding.GetByteCount(Input: Chinese)

```log

```


### Docs

[Profiling workflow for dotnet/runtime repository](https://github.com/dotnet/performance/blob/master/docs/profiling-workflow-dotnet-runtime.md)
[Benchmarking workflow for dotnet/runtime repository](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-dotnet-runtime.md)

</details>